Updated StackArray.py
Updated the Push function in the file.
Append function of lists was used to insert the key element in the stack which was wrong as append function  adds the key element at the end of the current list .So inspite of adding element to the index pointed by top the code was adding element to the end of the list.

P.S. You could verify this with the commented code.